requirements.txt
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
opencv-python>=4.8.0
Pillow>=10.0.0
PyAudio>=0.2.13
keyboard>=0.13.5



pip install google-genai
